Add support for http listener and improvement performance
